Prevent excessive updates on key press, extract useKeyboard, useMouse hooks
