+++ src/Spreadsheet/index.d.ts
@@ -1,6 +1,6 @@
-import { MutableRefObject, HTMLAttributes, Dispatch, SetStateAction, FunctionComponent, MouseEventHandler, Context, CSSProperties, DOMElement } from 'react';
-import { ScrollerCellProps } from '../Scroller';
-import { MergedCellProps } from '../MergedCell';
+import { MutableRefObject, HTMLAttributes, Dispatch, SetStateAction, FunctionComponent, MouseEventHandler, Context, CSSProperties } from 'react';
+import { ScrollerCellProps } from '../Scroller/index';
+import { MergedCellProps } from '../MergedCell/index';
 
 export interface Font {
   name?: string;
@@ -148,8 +148,6 @@ export const SpreadsheetContainer: FunctionComponent<SpreadsheetContainerProps>
 export interface UseSpreadsheetOptions {
   cells?: Cell[][];
   onCellsChange?: Dispatch<SetStateAction<Cell[][]>>;
-  defaultRowHeight: number;
-  defaultColumnWidth: number;
   rows?: Meta[]; 
   columns?: Meta[];
   onRowsChange?: Dispatch<SetStateAction<Meta[]>>;
@@ -168,8 +166,6 @@ export interface UseSpreadsheetOptions {
    */
   groupSize?: number;
   mergedCells?: CellsRange[];
-  rowsPerPage: number;
-  columnsPerPage: number;
   totalRows: number;
   totalColumns: number;
   fixRows?: number;
@@ -207,9 +203,10 @@ export interface UseSpreadsheetResult {
   scrollerContainerRef: MutableRefObject<Element>;
   scrollerCoverRef: MutableRefObject<Element>;
   spreadsheetContainerRef: MutableRefObject<Element>;
-  onKeyDown: KeyboardEvent;
 }
 export function useSpreadsheet(options: UseSpreadsheetOptions): UseSpreadsheetResult
+export function useKeyboard(options: UseSpreadsheetOptions | UseSpreadsheetResult): KeyboardEvent
+export function useMouse(options: UseSpreadsheetOptions | UseSpreadsheetResult): UIEvent
 
 export const SpreadsheetCell: FunctionComponent<ScrollerCellProps | MergedCellProps>
