+++ src/Spreadsheet/index.js
@@ -2,4 +2,6 @@ export { default as SpreadsheetCell } from './SpreadsheetCell';
 export { default as SpreadsheetContext } from './SpreadsheetContext';
 export { default as SpreadsheetContainer } from './SpreadsheetContainer';
 export { default as useSpreadsheet } from './useSpreadsheet';
+export { default as useKeyboard } from './useKeyboard';
+export { default as useMouse } from './useMouse';
 export { default } from './Spreadsheet';
